+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added MRMS data source
 - Added IFS and IFS-ENS initial state datasources
 - Local datasource example
+- Added ``TimeWindow`` datasource wrapper for fetching data at multiple time offsets
+- Added NOAA's Integrated Surface Database (ISD) data frame source
 
 ### Changed
 
@@ -34,6 +36,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - IMERG datasource
 
 ### Fixed
+
+- Vertical wind variables in GFS / GEFS to be Pa s-1
 
 ### Security